- Added unit tests of CurrencyExchangerService - Added UML Diagram - Added validations

+++ src/main/java/com/currencyexchanger/domain/response/ExchangeRateApiResponseDTO.java
@@ -19,5 +19,5 @@ public class ExchangeRateApiResponseDTO {
     @JsonProperty("target_code")
     private String targetCode;
     @JsonProperty("conversion_rate")
-    private float conversionRate;
+    private double conversionRate;
 }
